Feature2/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
python-dotenv
google-generativeai
faiss-cpu
pydantic